+++ packages/flame/test/components/tapables_test.dart
@@ -2,25 +2,25 @@ import 'package:flame/components.dart';
 import 'package:flame/game.dart';
 import 'package:test/test.dart';
 
-class _GameWithTapables extends BaseGame with HasTapableComponents {}
+class _GameWithTappables extends BaseGame with HasTappableComponents {}
 
-class _GameWithoutTapables extends BaseGame {}
+class _GameWithoutTappables extends BaseGame {}
 
-class TapableComponent extends PositionComponent with Tapable {}
+class TappableComponent extends PositionComponent with Tappable {}
 
 void main() {
-  group('tapables test', () {
+  group('tappables test', () {
     test('make sure they cannot be added to invalid games', () async {
-      final game1 = _GameWithTapables();
+      final game1 = _GameWithTappables();
       game1.onResize(Vector2.all(100));
       // should be ok
-      await game1.add(TapableComponent());
+      await game1.add(TappableComponent());
 
-      final game2 = _GameWithoutTapables();
+      final game2 = _GameWithoutTappables();
       game2.onResize(Vector2.all(100));
 
       expect(
-        () => game2.add(TapableComponent()),
+        () => game2.add(TappableComponent()),
         throwsA(isA<AssertionError>()),
       );
     });